Provides an interface to an email connector.
Define a url for an email connector that will connect to an email list.
The url can be a server:port, program, etc.
Name (First, Last) email address, and connector password will be sent to the connector.

The purpose of email connectors is to provide autonomy in the choice
of email services and programs.  E.g. mailman, googlegroups
See ./examples for an example connector