$0 deposits now acceptable for record-keeping reasons
Originally, deposits of $0 (amount > 0) would not be considered real deposits, however, there may be shops where only non-monetary transactions occurred (amount >= 0) which would be useful to record in an accounting program.
One caveat, if a monetary transaction is recorded, but the depositor only enters $0, the deposit will show "Difference: n/a", however this should be a cue since it should be obvious that a real world deposit of $0 would not be made at a bank.